+++ lib-e131/src/node/e131params.cpp
@@ -2,7 +2,7 @@
  * @file e131params.cpp
  *
  */
-/* Copyright (C) 2016-2023 by Arjan van Vught mailto:info@orangepi-dmx.nl
+/* Copyright (C) 2016-2024 by Arjan van Vught mailto:info@orangepi-dmx.nl
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -387,7 +387,7 @@ void E131Params::Dump() {
 
 	for (uint32_t i = 0; i < e131params::MAX_PORTS; i++) {
 		const auto nOutputStyle = static_cast<uint32_t>(isOutputStyleSet(1U << i));
-		printf(" %s=%u [%s]\n", LightSetParamsConst::OUTPUT_STYLE[i], nOutputStyle, lightset::get_output_style(static_cast<lightset::OutputStyle>(nOutputStyle)));
+		printf(" %s=%u [%s]\n", LightSetParamsConst::OUTPUT_STYLE[i], static_cast<unsigned int>(nOutputStyle), lightset::get_output_style(static_cast<lightset::OutputStyle>(nOutputStyle)));
 	}
 
 	if (isMaskSet(e131params::Mask::DISABLE_MERGE_TIMEOUT)) {
