+++ lib-rdmsubdevice/src/rdmsubdevicesparams.cpp
@@ -2,7 +2,7 @@
  * @file rdmsubdevicesparams.cpp
  *
  */
-/* Copyright (C) 2020-2023 by Arjan van Vught mailto:info@gd32-dmx.org
+/* Copyright (C) 2020-2024 by Arjan van Vught mailto:info@gd32-dmx.org
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -134,7 +134,7 @@ void RDMSubDevicesParams::Builder(const rdm::subdevicesparams::Params *pParams, 
 	for (uint32_t nCount = 0; nCount < m_Params.nCount; nCount++) {
 		char buffer[32];
 		const auto *p = &m_Params.Entry[nCount];
-		snprintf(buffer, sizeof(buffer) - 1, "%u,%s,%u,%u,%u", p->nChipSelect, rdm::subdevices::get_type_string(static_cast<rdm::subdevices::Types>(p->nType)), p->nAddress, p->nDmxStartAddress, p->nSpeedHz);
+		snprintf(buffer, sizeof(buffer) - 1, "%u,%s,%u,%u,%u", p->nChipSelect, rdm::subdevices::get_type_string(static_cast<rdm::subdevices::Types>(p->nType)), p->nAddress, p->nDmxStartAddress, static_cast<unsigned int>(p->nSpeedHz));
 		builder.AddRaw(buffer);
 	}
